refactor(mina-consensus.ts, utils.ts): move findAllIndexes function from mina-consensus.ts to utils.ts for better code organization and reusability
